Improve existing PostgreSQL schema to use better data types. Closes #18

+++ springboot-note/src/main/resources/db/migration/V1.0__CreateSchema.sql
@@ -1,7 +1,7 @@
 ----- STICKY NOTES
 
 CREATE TABLE sticky_note (
-    id NUMERIC(20) PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     title VARCHAR(1024) NOT NULL,
     body VARCHAR(10000) NOT NULL,
     type VARCHAR(32) NOT NULL,
@@ -12,8 +12,8 @@ CREATE INDEX sticky_note_title_idx ON sticky_note(title);
 
 
 CREATE TABLE sticky_note_link (
-    id NUMERIC(20) PRIMARY KEY,
-    sticky_note_id NUMERIC(20) NOT NULL,
+    id BIGINT PRIMARY KEY,
+    sticky_note_id BIGINT NOT NULL,
     link VARCHAR(1024) NOT NULL,
     CONSTRAINT sticky_note_link_sticky_note_id_fk FOREIGN KEY (sticky_note_id) REFERENCES sticky_note(id)
 );
@@ -23,7 +23,7 @@ CREATE INDEX sticky_note_link_sticky_note_id_idx ON sticky_note_link(sticky_note
 ----- AUTHENTICATION
 
 CREATE TABLE users (
-    id NUMERIC(20) PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     username VARCHAR(256) UNIQUE NOT NULL,
     password VARCHAR(256) NOT NULL,
     enabled NUMERIC(1) NOT NULL,
@@ -36,7 +36,7 @@ CREATE UNIQUE INDEX users_username_idx ON users(username);
 
 
 CREATE TABLE roles (
-    id NUMERIC(20) PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     rolename VARCHAR(256) UNIQUE NOT NULL
 );
 CREATE SEQUENCE roles_seq  START WITH 1 INCREMENT BY 1;
@@ -44,9 +44,9 @@ CREATE UNIQUE INDEX roles_rolename_idx ON roles(rolename);
 
 
 CREATE TABLE user_roles (
-    id NUMERIC(20) PRIMARY KEY,
-    user_id NUMERIC(20) NOT NULL,
-    role_id NUMERIC(20) NOT NULL,
+    id BIGINT PRIMARY KEY,
+    user_id BIGINT NOT NULL,
+    role_id BIGINT NOT NULL,
     CONSTRAINT user_roles_user_id_fk FOREIGN KEY (user_id) REFERENCES users(id),
     CONSTRAINT user_roles_role_id_fk FOREIGN KEY (role_id) REFERENCES roles(id)
 );
